+++ core/src/main/scala/japgolly/scalajs/react/vdom/ReactVDom.scala
@@ -56,6 +56,15 @@ object ReactVDom
 
     override type Tag = ReactVDom.Tag
 
+    private def genericJsAttr[T <% js.Any]: GenericAttr[T] = new GenericAttr[T](a => a)
+    override implicit val booleanAttr = genericJsAttr[Boolean]
+    override implicit val byteAttr    = genericJsAttr[Byte]
+    override implicit val shortAttr   = genericJsAttr[Short]
+    override implicit val intAttr     = genericJsAttr[Int]
+    override implicit val longAttr    = genericJsAttr[Long]
+    override implicit val floatAttr   = genericJsAttr[Float]
+    override implicit val doubleAttr  = genericJsAttr[Double]
+
     implicit val jsThisFnAttr = new GenericAttr[js.ThisFunction](f => f)
     implicit val jsFnAttr = new GenericAttr[js.Function](f => f)
     implicit def reactRefAttr[T <: Ref[_]] = new GenericAttr[T](_.name)